Fix inconsistent mapping in image plugin
'size' is inconsistent with volume plugin, 'status' and 'visibility' is
inconsistent with metadef plugin.

Closes-bug: #1570199
(cherry picked from commit 5db561268b1ab1f157862911bf3351d0423e9343)

+++ searchlight/elasticsearch/plugins/cinder/volumes.py
@@ -97,7 +97,8 @@ class VolumeIndex(base.IndexBase):
                         'raw': {'type': 'string', 'index': 'not_analyzed'}
                     }
                 },
-                'size': {'type': 'integer'},
+                # Using long here to consistent with glance.images
+                'size': {'type': 'long'},
                 'snapshot_id': {'type': 'string', 'index': 'not_analyzed'},
                 'source_volid': {'type': 'string', 'index': 'not_analyzed'},
                 'status': {'type': 'string', 'index': 'not_analyzed'},
